ARBattle
This work aims to present a study of collectible
card game mechanics interconnected with augmented
reality, an overview about the concepts of computer
vision and augmented reality. This article* also presents
works that use the augmented reality in games. It is
seen in an introductory way the collectible card games
origins. After that, it's described the development of
the game for this paper, followed by the explanation of
your build environment, describing tools and software
used to create it. Then, are told how operating
structure. Finally, is structured the obtained results and
the next steps to complete the development of this
game.

*Complete article download:
http://sbgames.org/sbgames2012/proceedings/papers/computacao/comp-short_05.pdf

Authors: Thiago Malheiros Porcino, Leandro Reis de Barros and John Edson
de Carvalho